+++ src/pipelines/sq-2025/pipeline/results/helpers/create-empty-results.js
@@ -21,5 +21,7 @@ export function createEmptyResults() {
     projectLinesOfCode: [],
     environment: null,
     configuration: null,
+    rulesComparisonData: null,
+    issuesDeltaData: null,
   };
 }