+++ src/pipelines/sq-2025/sonarcloud/reports/issues-delta/index.js
@@ -1,0 +1,1 @@
+export { gatherAllDelta } from './helpers/gather-all-delta.js';